pyproject.toml
[project]
name = "python-criando-agentes-com-agnos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agent>=0.1.3",
    "agno>=2.3.4",
    "groq>=0.36.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
    "yfinance>=0.2.66",
]
